// src/app/theme/shared/_helpers/models/Navigation.model.ts
export interface NavigationBadge {
  title?: string;
  type?: string;
}

export type NavigationType = 'item' | 'collapse' | 'group';

export interface NavigationItem {
  id: string;
  title: string;
  type: 'item' | 'collapse' | 'group';
  translate?: string;
  icon?: string;
  hidden?: boolean;
  url?: string;
  classes?: string;
  exactMatch?: boolean;
  external?: boolean;
  target?: boolean;
  breadcrumbs?: boolean;
  badge?: {
    title?: string;
    type?: string;
  };
  children?: NavigationItem[];
}